requirements.txt
# Qwen Code-Output Alignment Pipeline
# Install: pip install -r requirements.txt

# Core training
torch>=2.3.0
transformers>=4.45.0
datasets>=2.19.0
accelerate>=0.30.0

# TRL — contains SFTTrainer, DPOTrainer, GRPOTrainer
trl>=0.9.0

# LoRA / QLoRA
peft>=0.11.0
bitsandbytes>=0.43.0   # 4-bit quantization

# Utilities
scipy
numpy
tqdm
pyyaml

# Optional: experiment tracking
# wandb

# Optional: flash attention (significant speedup on A100/H100)
# flash-attn --no-build-isolation
